requirements.txt

chromadb==1.0.21
google_api_python_client==2.181.0
llama_index==0.14.2
pandas==2.3.2
PyPDF2==3.0.1
python-dotenv==1.1.1
streamlit==1.49.1
tqdm==4.67.1
unstructured==0.7.5



